move: hooks 폴더 위치 변경
